hack_backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
dj-database-url==2.0.0
Django==4.2.17
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.12.2
